chatbot_api/requirements.txt
opencv-python
joblib
google-generativeai
chromadb==0.5.0
streamlit==1.34.0
python-dotenv